@@ -6,44 +6,45 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
     "@babel/core": "^7.0.0",
     "@emotion/babel-plugin": "^11.10.2",
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^5.16.6",
-    "@mui/material": "^5.16.6",
+    "@eslint/migrate-config": "^1.6.0",
+    "@mui/icons-material": "^7.3.3",
+    "@mui/material": "^7.3.3",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@types/react-slick": "^0.23.10",
+    "@types/react-slick": "^0.23.13",
     "dayjs": "^1.11.5",
-    "destyle.css": "^3.0.2",
-    "i18next": "^23.11.2",
-    "next": "^14.0.4",
-    "next-i18next": "^15.3.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "destyle.css": "^4.0.1",
+    "next": "^15.5.4",
+    "next-intl": "^4.3.9",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
     "react-hook-form": "^7.48.2",
-    "react-i18next": "^14.1.0",
     "react-intersection-observer": "^9.4.4",
-    "react-slick": "^0.29.0",
-    "sharp": "^0.31.0",
+    "react-slick": "^0.31.0",
+    "sharp": "^0.34.4",
     "slick-carousel": "^1.8.1"
   },
   "devDependencies": {
-    "@types/node": "^20.11.4",
-    "@types/nodemailer": "^6.4.14",
-    "@types/react": "^18.2.48",
-    "@types/react-dom": "^18.2.18",
+    "@eslint/eslintrc": "^3",
+    "@tailwindcss/postcss": "^4.1.14",
+    "@types/node": "^24.6.2",
+    "@types/nodemailer": "^7.0.2",
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
     "autoprefixer": "^10.4.12",
-    "eslint": "8.24.0",
-    "eslint-config-next": "12.3.1",
+    "eslint": "9.36.0",
+    "eslint-config-next": "15.5.4",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-unused-imports": "^2.0.0",
-    "nodemailer": "^6.9.7",
+    "eslint-plugin-unused-imports": "^4.2.0",
+    "nodemailer": "^7.0.6",
     "postcss": "^8.4.18",
-    "tailwindcss": "^3.2.1",
-    "typescript": "4.8.3"
+    "tailwindcss": "^4.1.14",
+    "typescript": "~5.9.3"
   }
 }
