@@ -1,10 +1,8 @@
-import Link from "next/link";
+import { Link } from "@/i18n/routing";
 
-import Layout from "../../components/layouts/Layout";
-
-const Success = () => {
+export default function ContactSuccessPage() {
   return (
-    <Layout title="送信完了">
+    <>
       <div className="w-full h-screen flex justify-center items-center">
         <div className="flex flex-col justify-center items-center gap-4">
           <div className="w-32 h-32">
@@ -19,14 +17,12 @@ const Success = () => {
             <p className="text-center text-sm mb-2">
               担当者より返信を差し上げますのでお待ち下さい
             </p>
-            <Link href="/" className="text-ares-red ease-in-out">
+            <Link href="/" className="textaresred ease-in-out">
               トップページへ戻る
             </Link>
           </div>
         </div>
       </div>
-    </Layout>
+    </>
   );
-};
-
-export default Success;
+}
