@@ -1,15 +1,19 @@
-import { NextPage } from "next";
-import Link from "next/link";
+"use client";
 
-import contents from "../../const/header";
+import contents from "@/const/header";
+import { Link } from "@/i18n/routing";
 
 type Props = {
   open: boolean;
 };
 
-const Drawer: NextPage<Props> = ({ open }: Props) => {
+const Drawer = ({ open }: Props) => {
   return (
-    <nav className={`${open ? "block" : "hidden"} md:hidden fixed w-screen h-screen top-0 right-0`}>
+    <nav
+      className={`${
+        open ? "block" : "hidden"
+      } md:hidden fixed w-screen h-screen top-0 right-0`}
+    >
       <div className="flex w-full h-full flex-col pt-16">
         <div className="absolute w-full h-full top-0 bg-black opacity-50"></div>
         <ul className="md:hidden flex flex-col items-center gap-10 text-m font-display pt-10 z-10">
