@@ -1,6 +1,7 @@
+"use client";
 
-import LanguageIcon from "@mui/icons-material/Language";
-import Link from "next/link";
+import { Link, usePathname } from "@/i18n/routing";
+import { LanguageIcon } from "@/lib/mui";
 
 export const LocaleSwitcher = ({
   href,
@@ -9,28 +10,20 @@ export const LocaleSwitcher = ({
   href: string;
   locale: string;
 }) => {
+  const pathname = usePathname();
+  const targetLocale: "en" | "ja" = locale === "en" ? "ja" : "en";
+  const currentHref = pathname ?? href ?? "/";
+
   return (
     <div className="flex items-center gap-2 text-xs">
-      {locale === "en" && (
-        <Link
-          href={href}
-          locale="ja"
-          className="cursor-pointer hover:text-red-600 duration-300 flex items-center text-white"
-        >
-          <LanguageIcon className="h-4" />
-          <span className="">JA</span>
-        </Link>
-      )}
-      {locale === "ja" && (
-        <Link
-          href={href}
-          locale="en"
-          className="cursor-pointer hover:text-red-600 duration-300 flex items-center text-white"
-        >
-          <LanguageIcon className="h-4" />
-          <span className="align-middle">EN</span>
-        </Link>
-      )}
+      <Link
+        href={currentHref}
+        locale={targetLocale}
+        className="cursor-pointer hover:text-red-600 duration-300 flex items-center text-white"
+      >
+        <LanguageIcon className="h-4" />
+        <span className="align-middle">{targetLocale.toUpperCase()}</span>
+      </Link>
     </div>
   );
 };
