@@ -1,11 +1,14 @@
-import type { NextPage } from "next";
-import { useTranslation } from "next-i18next";
+"use client";
+import dynamic from "next/dynamic";
+import { useTranslations } from "next-intl";
 
-import { mainImages } from "../../../const/images";
-import Slide from "../atoms/Slide";
+const Slide = dynamic(() => import("@/components/objects/atoms/Slide"), {
+  ssr: false,
+});
+import { mainImages } from "@/const/images";
 
-const MainVisual: NextPage = () => {
-  const { t } = useTranslation("common");
+const MainVisual = () => {
+  const t = useTranslations();
   return (
     <div className="relative h-screen">
       <div className="h-16 md:h-0"></div>
@@ -16,9 +19,9 @@ const MainVisual: NextPage = () => {
           className: "absolute top-0 left-0 w-full h-full overflow-hidden",
         }}
       />
-      <div className="absolute top-0 left-0 w-full h-full bg-main-bg/75">
+      <div className="absolute top-0 left-0 w-full h-full bgmainbg75">
         <div className="w-full h-full flex flex-col justify-end gap-2 pl-2 md:pl-20 pb-[30vh] md:pb-20">
-          <h2 className=" font-display text-4xl md:text-6xl animate-[emerge1_3s_ease-out] tracking-wider">
+          <h2 className="font-display text-4xl md:text-6xl animate-[emerge1_3s_ease-out] tracking-wider">
             A<span className="text-red-600">R</span>ES Project
           </h2>
           <p className="text-xl md:text-3xl animate-[emerge2_3s_ease-out]">
